test/regression/reftest/web-platform-tests/dom/nodes/Document-createTextNode.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.createTextNode</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-document-createtextnode">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-ownerdocument">
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-data">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-nodevalue">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-textcontent">
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-length">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-nodetype">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-haschildnodes">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-childnodes">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-firstchild">
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-lastchild">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<!--script src="Document-createComment-createTextNode.js"></script-->
<div id="log"></div>
<script>
function test_create(method, iface, nodeType, nodeName) {
  ["\u000b", "a -- b", "a-", "-b", null, undefined].forEach(function(value) {
    test(function() {
      var c = document[method](value);
      var expected = String(value);
      assert_true(c instanceof iface);
      assert_true(c instanceof CharacterData);
      assert_true(c instanceof Node);
      assert_equals(c.ownerDocument, document);
      assert_equals(c.data, expected, "data");
      assert_equals(c.nodeValue, expected, "nodeValue");
      assert_equals(c.textContent, expected, "textContent");
      assert_equals(c.length, expected.length);
      assert_equals(c.nodeType, nodeType);
      assert_equals(c.nodeName, nodeName);
      assert_equals(c.hasChildNodes(), false);
      assert_equals(c.childNodes.length, 0);
      assert_equals(c.firstChild, null);
      assert_equals(c.lastChild, null);
    }, method + "(" + format_value(value) + ")");
  });
}

test_create("createTextNode", Text, 3, "#text");
</script>
